+++ spi/transfer-spi/src/main/java/org/eclipse/dataspaceconnector/spi/transfer/store/TransferProcessStore.java
@@ -14,20 +14,19 @@
 
 package org.eclipse.dataspaceconnector.spi.transfer.store;
 
+import org.eclipse.dataspaceconnector.spi.persistence.StateEntityStore;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
 import org.eclipse.dataspaceconnector.spi.system.Feature;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
 import java.util.stream.Stream;
 
 /**
  * Manages persistent storage of {@link TransferProcess} state.
  */
 @Feature(TransferProcessStore.FEATURE)
-public interface TransferProcessStore {
+public interface TransferProcessStore extends StateEntityStore<TransferProcess> {
     String FEATURE = "edc:core:transfer:transferprocessstore";
 
     /**
@@ -41,23 +40,6 @@ public interface TransferProcessStore {
      */
     @Nullable
     String processIdForTransferId(String id);
-
-    /**
-     * Returns a list of TransferProcesses that are in a specific state.
-     * <br/>
-     * Implementors MUST handle the starvation scenario, i.e. when the number of processes is greater than the number
-     * passedin via {@code max}.
-     * E.g. database-based implementations should perform a query along the lines of {@code SELECT ... ORDER BY TransferProcess#stateTimestamp}.
-     * Then, after the check, users of this method must update the {@code TransferProcess#stateTimestamp} even if the process
-     * remains unchanged.
-     * Some database frameworks such as Spring have automatic lastChanged columns.
-     *
-     * @param state The state that the processes of interest should be in.
-     * @param max   The maximum amount of result items.
-     * @return A list of TransferProcesses (at most _max_) that are in the desired state.
-     */
-    @NotNull
-    List<TransferProcess> nextForState(int state, int max);
 
     /**
      * Creates a transfer process.